+++ posthog/src/main/java/com/posthog/PostHog.kt
@@ -209,15 +209,10 @@ public class PostHog private constructor(
 
                 // only because of testing in isolation, this flag is always enabled
                 if (reloadFeatureFlags) {
-                    when {
-                        config.remoteConfig ->
-                            loadRemoteConfigRequest(
-                                internalOnFeatureFlagsLoaded,
-                                config.onFeatureFlags,
-                            )
-
-                        config.preloadFeatureFlags -> reloadFeatureFlags(config.onFeatureFlags)
-                    }
+                    loadRemoteConfigRequest(
+                        internalOnFeatureFlagsLoaded,
+                        config.onFeatureFlags,
+                    )
                 }
             } catch (e: Throwable) {
                 config.logger.log("Setup failed: $e.")