+++ posthog/src/test/java/com/posthog/PostHogFeatureFlagsTest.kt
@@ -41,7 +41,6 @@ internal class PostHogFeatureFlagsTest {
         sendFeatureFlagEvent: Boolean = true,
         reuseAnonymousId: Boolean = false,
         integration: PostHogIntegration? = null,
-        remoteConfig: Boolean = false,
         cachePreferences: PostHogMemoryPreferences = PostHogMemoryPreferences(),
         propertiesSanitizer: PostHogPropertiesSanitizer? = null,
         beforeSend: PostHogBeforeSend? = null,
@@ -64,7 +63,6 @@ internal class PostHogFeatureFlagsTest {
                 this.cachePreferences = cachePreferences
                 this.propertiesSanitizer = propertiesSanitizer
                 this.evaluationContexts = evaluationContexts
-                this.remoteConfig = remoteConfig
                 if (beforeSend != null) {
                     addBeforeSend(beforeSend)
                 }
@@ -102,7 +100,7 @@ internal class PostHogFeatureFlagsTest {
                 .setBody(""),
         )
         val url = http.url("/")
-        val sut = getSut(url.toString(), preloadFeatureFlags = false, sendFeatureFlagEvent = false)
+        val sut = getSut(url.toString(), preloadFeatureFlags = false, sendFeatureFlagEvent = false, reloadFeatureFlags = false)
 
         sut.reloadFeatureFlags()
 
@@ -141,7 +139,7 @@ internal class PostHogFeatureFlagsTest {
                 .setBody(""),
         )
         val url = http.url("/")
-        val sut = getSut(url.toString(), preloadFeatureFlags = false, sendFeatureFlagEvent = false)
+        val sut = getSut(url.toString(), preloadFeatureFlags = false, sendFeatureFlagEvent = false, reloadFeatureFlags = false)
 
         sut.reloadFeatureFlags()
         remoteConfigExecutor.shutdownAndAwaitTermination()
@@ -184,6 +182,7 @@ internal class PostHogFeatureFlagsTest {
                 url.toString(),
                 preloadFeatureFlags = false,
                 sendFeatureFlagEvent = false,
+                reloadFeatureFlags = false,
             )
 
         sut.reloadFeatureFlags()
@@ -224,7 +223,7 @@ internal class PostHogFeatureFlagsTest {
                 .setBody(""),
         )
         val url = http.url("/")
-        val sut = getSut(url.toString(), preloadFeatureFlags = false, sendFeatureFlagEvent = true)
+        val sut = getSut(url.toString(), preloadFeatureFlags = false, sendFeatureFlagEvent = true, reloadFeatureFlags = false)
 
         sut.reloadFeatureFlags()
 
@@ -264,7 +263,7 @@ internal class PostHogFeatureFlagsTest {
                 .setBody(""),
         )
         val url = http.url("/")
-        val sut = getSut(url.toString(), preloadFeatureFlags = false, sendFeatureFlagEvent = true)
+        val sut = getSut(url.toString(), preloadFeatureFlags = false, sendFeatureFlagEvent = true, reloadFeatureFlags = false)
 
         sut.reloadFeatureFlags()
 
@@ -307,6 +306,7 @@ internal class PostHogFeatureFlagsTest {
                 url.toString(),
                 preloadFeatureFlags = false,
                 sendFeatureFlagEvent = true,
+                reloadFeatureFlags = false,
             )
 
         sut.reloadFeatureFlags()
@@ -346,7 +346,7 @@ internal class PostHogFeatureFlagsTest {
                 .setBody(""),
         )
         val url = http.url("/")
-        val sut = getSut(url.toString(), preloadFeatureFlags = false, sendFeatureFlagEvent = true)
+        val sut = getSut(url.toString(), preloadFeatureFlags = false, sendFeatureFlagEvent = true, reloadFeatureFlags = false)
 
         sut.reloadFeatureFlags()